Rework key and salt generation.
Key generation is now the responsibility of a paste, because it knows
its own salt. Salts are now randomly-generated strings of base32,
instead of the paste ID. Encrypted paste IDs are now 8 characters. Paste
ID generation is now the responsibility of the paste store.

I'm not completely comfortable with how this turned out and might have
to revisit it later.