Language UI overhaul.
Fetch languages as JSON from the server, which is rendered when
languages are loaded.
Use select2's various data and display functions to override matching so
that we match on both the language title and the language's id.
Auxiliary IDs are matched as well. This drops the size of our homepage
itself (json notwithstanding) significantly.

The select element had to become an <input type="hidden">, but that
doesn't seem like too big a deal.

Javascript is now fully required.

Omit fields that are unnecessary in the resulting JSON.